+++ ee/clickhouse/queries/funnels/funnel_unordered.py
@@ -27,10 +27,12 @@ class ClickhouseFunnelUnordered(ClickhouseFunnelBase):
 
         max_steps = len(self._filter.entities)
 
+        breakdown_clause = self._get_breakdown_prop()
+
         return f"""
-        SELECT {self._get_count_columns(max_steps)} {self._get_step_time_avgs(max_steps)} FROM (
+        SELECT {self._get_count_columns(max_steps)} {self._get_step_time_avgs(max_steps)} {breakdown_clause} FROM (
             {self.get_step_counts_query()}
-        ) SETTINGS allow_experimental_window_functions = 1
+        ) {'GROUP BY prop' if breakdown_clause != '' else ''} SETTINGS allow_experimental_window_functions = 1
         """
 
     def get_step_counts_query(self):
@@ -38,13 +40,14 @@ class ClickhouseFunnelUnordered(ClickhouseFunnelBase):
         max_steps = len(self._filter.entities)
 
         union_query = self.get_step_counts_without_aggregation_query()
+        breakdown_clause = self._get_breakdown_prop()
 
         return f"""
-            SELECT person_id, steps {self._get_step_time_avgs(max_steps)} FROM (
-                SELECT person_id, steps, max(steps) over (PARTITION BY person_id) as max_steps {self._get_step_time_names(max_steps)} FROM (
+            SELECT person_id, steps {self._get_step_time_avgs(max_steps)} {breakdown_clause} FROM (
+                SELECT person_id, steps, max(steps) over (PARTITION BY person_id {breakdown_clause}) as max_steps {self._get_step_time_names(max_steps)} {breakdown_clause} FROM (
                         {union_query}
                 )
-            ) GROUP BY person_id, steps
+            ) GROUP BY person_id, steps {breakdown_clause}
             HAVING steps = max_steps
         """
 
@@ -55,6 +58,7 @@ class ClickhouseFunnelUnordered(ClickhouseFunnelBase):
 
         partition_select = self._get_partition_cols(1, max_steps)
         sorting_condition = self.get_sorting_condition(max_steps)
+        breakdown_clause = self._get_breakdown_prop()
 
         for i in range(max_steps):
             inner_query = f"""
@@ -62,6 +66,7 @@ class ClickhouseFunnelUnordered(ClickhouseFunnelBase):
                 person_id,
                 timestamp,
                 {partition_select}
+                {breakdown_clause}
                 FROM ({self._get_inner_event_query(entities_to_use, f"events_{i}")})
             """
